requirements.txt
# Core dependencies
streamlit>=1.31.0
pandas>=2.2.0
openai>=1.12.0
python-dotenv>=1.0.0

#data base
gspread==5.12.4
google-auth==2.28.1

# Development and file watching
watchdog>=3.0.0

# Date handling
python-dateutil>=2.8.2

# Type checking
typing-extensions>=4.5.0

# Path handling
pathlib>=1.0.1

# Data handling
uuid>=1.30

# Optional but recommended for better performance
streamlit-extras>=0.3.0  # Additional Streamlit components